fix: adapt restart --interactive to current cli.rs structure, split test
